Moved <ObjHeader> into objects directory
