+++ media/src/objects/Point.svelte
@@ -14,7 +14,7 @@
 
     import PlayButtons from '../form-components/PlayButtons.svelte';
     import M from '../M.svelte';
-    import ObjHeader from '../ObjHeader.svelte';
+    import ObjHeader from './ObjHeader.svelte';
     import { checksum } from '../utils.js';
     import InputChecker from '../form-components/InputChecker.svelte';
